pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hunyuan-image-mcp"
version = "0.2.0"
description = "Multi-API图像生成MCP服务器，支持stdio和HTTP传输"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "tencentcloud-sdk-python>=3.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=0.19.0",
    "mcp>=1.6.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
mcp-image-server = "mcp_image_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_image_server"] 